@nsajko 2024-10-11
I believe you want to target branch [backports-release-1.10](https://github.com/JuliaLang/julia/tree/backports-release-1.10) (PR #55746).